requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pytest>=7.0.0
redis>=5.0.0
prometheus-client>=0.16.0
requests>=2.0.0
